backend/requirements.txt
Django
dj-database-url
dj-static
django-appconf
django-compressor
djangorestframework
drf-nested-routers
gunicorn
six
static3
wsgiref
django-background-task
demjson
mysqlclient
